Clean up code

+++ src/main/kotlin/de/nihas101/imagesToPdfConverter/ProgressUpdater.kt
@@ -1,5 +1,5 @@
 package de.nihas101.imagesToPdfConverter
 
 interface ProgressUpdater{
-    fun updateProgress(progress: Int)
+    fun updateProgress(progress: Double)
 }
